More UI - Don't revert if tests don't work - READ DESC
Tests don't work due to bad structure when declaring module-info.

Tests are not part of the final .jar anyway, this makes the program runnable such that it can be turned into a jar. Source code will be without the module-info file since that still has the main code be runnable albeit with a negligible error.