A lot of UI elements - READ DESC
Still needs a lot of actual data retrieval and construction for this implementation.

Still missing "my activities page" as I don't think I have a way of seeing if the user is assigned to an activity or not(At least properly, i know i can iterate).

Also missing a template fxml for constructing tiles for both projects and activities, hence making them accessible and such.